requirements.txt
#For Development
ipykernel # for jupyter notebook
mypy # type checker
ruff # for python 2.7 python liter to higlight syntactical and stylistic problems in python code
black # code formatter

#Dependency Management
wandb # for tracking experiments
lightning
torchvision
pytorch-lightning[extra]